+++ lib/generatorlabs.rb
@@ -11,6 +11,8 @@
 
 require_relative 'generatorlabs/version'
 require_relative 'generatorlabs/config'
+require_relative 'generatorlabs/rate_limit_info'
+require_relative 'generatorlabs/response'
 require_relative 'generatorlabs/client'
 require_relative 'generatorlabs/request_handler'
 require_relative 'generatorlabs/webhook'